[Pedro Resende] Fluxo de contas

+++ src/main/java/com/pedroresende/rateiofacil/controllers/UserController.java
@@ -1,15 +1,18 @@
 package com.pedroresende.rateiofacil.controllers;
 
-import com.pedroresende.rateiofacil.controllers.dtos.CreationUserDto;
-import com.pedroresende.rateiofacil.controllers.dtos.ResponseDto;
-import com.pedroresende.rateiofacil.controllers.dtos.UserDto;
+import com.pedroresende.rateiofacil.controllers.dtos.*;
+import com.pedroresende.rateiofacil.exceptions.NotAuthorizeUserException;
+import com.pedroresende.rateiofacil.models.entities.Bill;
 import com.pedroresende.rateiofacil.models.entities.User;
 import com.pedroresende.rateiofacil.services.UserService;
+import com.pedroresende.rateiofacil.utils.Calculator;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.annotation.Secured;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -37,15 +40,12 @@ public class UserController {
    * Mapemaneto da rota POST /users.
    */
   @PostMapping()
-  public ResponseEntity<ResponseDto<UserDto>> create(
-      @RequestBody CreationUserDto creationUserDto) {
+  public ResponseEntity<ResponseDto<UserDto>> create(@RequestBody CreationUserDto creationUserDto) {
     User user = userService.create(creationUserDto.toEntity());
 
     UserDto userDto = toUserDto(user);
 
-    ResponseDto<UserDto> responseDto = new ResponseDto<>(
-        "Usuário criado com sucesso!", userDto
-    );
+    ResponseDto<UserDto> responseDto = new ResponseDto<>("Usuário criado com sucesso!", userDto);
 
     return ResponseEntity.status(HttpStatus.CREATED).body(responseDto);
   }
@@ -77,9 +77,11 @@ public class UserController {
    * Mapeamento da rota GET users/{id}.
    */
   @GetMapping("/{id}")
-  public ResponseEntity<UserDto> getById(@PathVariable Long id) {
+  public ResponseEntity<UserDto> getById(@PathVariable Long id,
+      @AuthenticationPrincipal User authUser) {
+    validateUserPermission(id, authUser);
     User user = userService.getById(id);
-
+    UserDto userDto = toUserDto(user);
     return ResponseEntity.ok(toUserDto(user));
   }
 
@@ -88,11 +90,12 @@ public class UserController {
    */
   @PutMapping("/{id}")
   public ResponseEntity<ResponseDto<UserDto>> update(@PathVariable Long id,
-      @RequestBody CreationUserDto creationUserDto) {
-    User user = userService.update(id, creationUserDto.toEntity());
-    ResponseDto<UserDto> responseDto = new ResponseDto<>(
-        "Usuário atualizado com sucesso", toUserDto(user)
-    );
+      @RequestBody UpdateUserDto updateUserDto,
+      @AuthenticationPrincipal User authUser) {
+    validateUserPermission(id, authUser);
+    User user = userService.update(id, updateUserDto.toEntity());
+    ResponseDto<UserDto> responseDto = new ResponseDto<>("Usuário atualizado com sucesso",
+        toUserDto(user));
 
     return ResponseEntity.ok(responseDto);
   }
@@ -101,11 +104,93 @@ public class UserController {
    * Mapeamento da rota DELETE users/{id}.
    */
   @DeleteMapping("/{id}")
-  public ResponseEntity<ResponseDto<UserDto>> delete(@PathVariable Long id) {
+  public ResponseEntity<ResponseDto<UserDto>> delete(@PathVariable Long id,
+      @AuthenticationPrincipal User authUser) {
+    validateUserPermission(id, authUser);
     User user = userService.delete(id);
     UserDto userDto = toUserDto(user);
     ResponseDto<UserDto> responseDto = new ResponseDto<>("Usuário deletado!", userDto);
 
     return ResponseEntity.ok(responseDto);
+  }
+
+  /**
+   * Mapeamento da rota POST /users/{id}/bills.
+   *
+   * @param id      identificador do usuário.
+   * @param billDto informações da conta.
+   * @return mensagem de sucesso e uma instancia de BillDto
+   */
+  @PostMapping("/{id}/bills")
+  @CrossOrigin("http://localhost:7173")
+  public ResponseEntity<ResponseDto<BillDto>> createBill(@PathVariable Long id,
+      @RequestBody BillDto billDto,
+      @AuthenticationPrincipal User authUser) {
+    validateUserPermission(id, authUser);
+    Bill bill = userService.associateBill(id, billDto.toEntity());
+
+    BillDto billDtoFromDb = toBillDto(bill);
+
+    ResponseDto<BillDto> responseDto = new ResponseDto<>("Conta associada com sucesso",
+        billDtoFromDb);
+    return ResponseEntity.status(HttpStatus.CREATED).body(responseDto);
+  }
+
+
+  private BillDto toBillDto(Bill bill) {
+
+    if (bill.getUser() == null) {
+      return new BillDto(bill.getId(), null, bill.getDate(), bill.getEstablishment(),
+          Calculator.sumValues(bill.getTotal(), 0.0), bill.getStatus());
+    }
+
+    return new BillDto(bill.getId(), bill.getUser().getId(), bill.getDate(),
+        bill.getEstablishment(), bill.getTotal(), bill.getStatus());
+  }
+
+  private BillDtoWithitems toBillDtoWithitems(Bill bill) {
+    if (bill.getItems() == null) {
+      return new BillDtoWithitems(bill.getId(), bill.getUser().getId(), bill.getDate(),
+          null, bill.getEstablishment(), bill.getTotal());
+    }
+    return new BillDtoWithitems(bill.getId(), bill.getUser().getId(), bill.getDate(),
+        bill.getItems(), bill.getEstablishment(), bill.getTotal());
+  }
+
+  @GetMapping("/{id}/bills")
+  ResponseEntity<List<BillDto>> getAllBills(@PathVariable Long id,
+      @AuthenticationPrincipal User authUser) {
+    validateUserPermission(id, authUser);
+    List<Bill> bills = userService.getAllBill(id);
+    List<BillDto> billDtos = bills.stream().map(bill -> toBillDto(bill)).toList();
+
+    return ResponseEntity.ok(billDtos);
+  }
+
+  @GetMapping("/{id}/bills/{billId}")
+  ResponseEntity<BillDto> getBill(@PathVariable Long id, @PathVariable Long billId,
+      @AuthenticationPrincipal User authUser) {
+    validateUserPermission(id, authUser);
+    Bill bill = userService.getBill(id, billId);
+
+    return ResponseEntity.ok(toBillDto(bill));
+  }
+
+  @DeleteMapping("/{id}/bills/{billId}")
+  ResponseEntity<ResponseDto<BillDto>> deleteBill(@PathVariable Long id, @PathVariable Long billId,
+      @AuthenticationPrincipal User authUser) {
+    validateUserPermission(id, authUser);
+    Bill bill = userService.deleteBill(id, billId);
+    BillDto billDto = toBillDto(bill);
+    ResponseDto<BillDto> responseDto = new ResponseDto<>("Conta deletada com sucesso", billDto);
+    return ResponseEntity.ok(responseDto);
+  }
+
+  private void validateUserPermission(Long pathId, User authUser) {
+    if (!authUser.getRole().equals("admin")) {
+      if (!authUser.getId().equals(pathId)) {
+        throw new NotAuthorizeUserException();
+      }
+    }
   }
 }
